+++ Classes/Emogrifier.php
@@ -159,7 +159,7 @@ class Emogrifier
      */
     private $xPathRules = [
         // child
-        '/\\s+>\\s+/'                              => '/',
+        '/\\s*>\\s*/'                              => '/',
         // adjacent sibling
         '/\\s+\\+\\s+/'                            => '/following-sibling::*[1]/self::',
         // descendant
@@ -174,6 +174,8 @@ class Emogrifier
         '/(\\w)\\[(\\w+)\\]/'                      => '\\1[@\\2]',
         // exact attribute
         '/(\\w)\\[(\\w+)\\=[\'"]?(\\w+)[\'"]?\\]/' => '\\1[@\\2="\\3"]',
+        // element attribute~=
+        '/([\\w\\*]+)\\[(\\w+)[\\s]*\\~\\=[\\s]*[\'"]?([\\w-_\\/]+)[\'"]?\\]/' => '\\1[contains(concat(" ", @\\2, " "), concat(" ", "\\3", " "))]',
     ];
 
     /**